+++ src/app/app.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {Footer, Body, TitleBar} from 'frontend/components'
+import { Footer, Body, TitleBar } from 'frontend/components'
 import autobind from 'autobind'
 
 interface State {
@@ -28,10 +28,16 @@ export default class App extends React.Component<{}, State> {
         this.updateZoom(this.state.zoomFactor - 0.1);
     }
 
+    resetZoom(e) {
+        e.preventDefault();
+        this.updateZoom(1.0);
+    }
+
     public render() {
         return (
             <div>
-                <TitleBar zoomIn={this.zoomIn} zoomOut={this.zoomOut} />
+                <TitleBar zoomIn={this.zoomIn} zoomOut={this.zoomOut} resetZoom={this.resetZoom}
+                    currZoom={Math.floor(this.state.zoomFactor * 100.0)} />
                 <Body zoomFactor={this.state.zoomFactor} />
                 <Footer />
             </div>