+++ src/app/frontend/components/body/LoadingPage.tsx
@@ -14,6 +14,7 @@ interface Props {
 
 export default class LoadingPage extends React.Component<Props, State> {
     timerID: NodeJS.Timer;
+
     constructor(props) {
         super(props)
         this.state = { dots: 0, progress: this.props.progress }
@@ -28,7 +29,26 @@ export default class LoadingPage extends React.Component<Props, State> {
         clearInterval(this.timerID)
     }
 
+    componentWillReceiveProps(props) {
+        this.setState({ progress: props.progress });
+    }
+
     public render() {
+        let res = this.state.progress
+        if (res < 1) {
+            res *= 100;
+        }
+        if(res < 0) {
+            res = 0;
+        }
+        res = Math.ceil(res)
+        console.log(res)
+        let progress
+        if (this.props.progress) {
+            progress = <rc_progress.Line percent={res} strokeWidth="2" trailWidth="2" className="progress" strokeColor="#FF52FF" />
+        } else {
+            progress = <rc_progress.Line percent={0} strokeWidth="2" trailWidth="2" className="progress" strokeColor="#FF52FF" />
+        }
         return (
             <div className="loading-container">
                 <div className="wrapper">
@@ -37,9 +57,7 @@ export default class LoadingPage extends React.Component<Props, State> {
                             {this.props.message}
                         </div>
                     }
-                    {this.props.progress &&
-                        <rc_progress.Line percent={this.state.progress * (this.state.progress < 1.01 ? 100 : 1)} strokeWidth="2" trailWidth="2" className="progress" strokeColor="#FF52FF" />
-                    }
+                    {progress}
                 </div>
             </div>
         )